+++ compiler/src/index.js
@@ -3,13 +3,10 @@ const fs = require("fs");
 const mysql = require("mysql");
 const psql = require("pg");
 const http = require("http");
-const Canvas = require("./Canvas").Canvas;
-const View = require("./View").View;
-const Jump = require("./Jump").Jump;
-const Layer = require("./Layer").Layer;
-const Transform = require("./Transform").Transform;
-const getBodyStringOfFunction = require("./template-api/Utilities")
-    .getBodyStringOfFunction;
+const addTable = require("./template-api/addTable").addTable;
+const addSSV = require("./template-api/addSSV").addSSV;
+const addUSMap = require("./template-api/addUSMap").addUSMap;
+const addPie = require("./template-api/addPie").addPie;
 
 /**
  *
@@ -180,584 +177,6 @@ function addJump(jump) {
     }
 
     this.jumps.push(jump);
-}
-
-// Add a Tabular vis to a project
-function addTable(table, args) {
-    if (args == null) args = {};
-
-    this.tables.push(table);
-    table.name = "kyrix_table_" + (this.tables.length - 1);
-
-    table.renderingParams = {
-        [table.name]: {
-            x: table.x,
-            y: table.y,
-            heads: {
-                height: table.heads_height,
-                names: table.heads_names
-            },
-            width: table.width,
-            cell_height: table.cell_height,
-            fields: table.schema.slice(0, table.schema.indexOf("rn"))
-        }
-    };
-
-    var canvas = new Canvas(
-        table.name,
-        Math.ceil(table.sum_width),
-        0,
-        "",
-        `0:select count(*) * ${table.cell_height} + ${table.heads_height} from ${table.table}`
-    );
-    this.addCanvas(canvas);
-    this.addStyles(__dirname + "/template-api/css/table.css");
-    this.addRenderingParams(table.renderingParams);
-    var transform_func = table.getTableTransformFunc();
-    var tableTransform = new Transform(
-        table.query,
-        table.db,
-        transform_func,
-        table.schema,
-        true
-    );
-
-    var tableLayer = new Layer(tableTransform, false);
-    tableLayer.addPlacement(table.placement);
-    tableLayer.addRenderingFunc(table.getTableRenderer());
-    if (table.group_by.length > 0) {
-        tableLayer.setIndexerType("PsqlPredicatedTableIndexer");
-    }
-    canvas.addLayer(tableLayer);
-
-    if (!args.view) {
-        var tableView = new View(
-            table.name + "_view",
-            Math.floor(table.sum_width * 0.8),
-            700
-        );
-        this.addView(tableView);
-        this.setInitialStates(tableView, canvas, 0, 0);
-    } else if (!(args.view instanceof View))
-        throw new Error("Adding Table: view must be a View object");
-
-    return {canvas, view: args.view ? args.view : tableView};
-}
-
-/**
- * Add an ssv to a project, this will create a hierarchy of canvases that form a pyramid shape
- * @param ssv an SSV object
- * @param args an dictionary that contains customization parameters, see doc
- */
-function addSSV(ssv, args) {
-    if (args == null) args = {};
-
-    // add to project
-    this.ssvs.push(ssv);
-
-    // add stuff to renderingParam
-    var renderingParams = {
-        textwrap: require("./template-api/Utilities").textwrap,
-        processClusterAgg: require("./template-api/SSV").processClusterAgg,
-        serializePath: require("./template-api/Utilities").serializePath,
-        translatePathSegments: require("./template-api/Utilities")
-            .translatePathSegments,
-        parsePathIntoSegments: require("./template-api/Utilities")
-            .parsePathIntoSegments,
-        aggKeyDelimiter: ssv.aggKeyDelimiter,
-        loX: ssv.loX,
-        loY: ssv.loY,
-        hiX: ssv.hiX,
-        hiY: ssv.hiY,
-        bboxW: ssv.bboxW,
-        bboxH: ssv.bboxH,
-        zoomFactor: ssv.zoomFactor,
-        fadeInDuration: 200,
-        geoInitialLevel: ssv.geoInitialLevel,
-        geoInitialCenterLat: ssv.geoLat,
-        geoInitialCenterLon: ssv.geoLon
-    };
-    renderingParams = {
-        ...renderingParams,
-        ...ssv.clusterParams,
-        ...ssv.aggregateParams,
-        ...ssv.hoverParams,
-        ...ssv.legendParams,
-        ...ssv.axisParams
-    };
-    var rpKey = "ssv_" + (this.ssvs.length - 1);
-    var rpDict = {};
-    rpDict[rpKey] = renderingParams;
-    this.addRenderingParams(rpDict);
-
-    // construct canvases
-    var curPyramid = [];
-    var transform = new Transform(ssv.query, ssv.db, "", [], true);
-    var numLevels = Math.min(
-        ssv.numLevels,
-        args.pyramid ? args.pyramid.length : 1e10
-    );
-    for (var i = 0; i < numLevels; i++) {
-        var width = (ssv.topLevelWidth * Math.pow(ssv.zoomFactor, i)) | 0;
-        var height = (ssv.topLevelHeight * Math.pow(ssv.zoomFactor, i)) | 0;
-
-        // construct a new canvas
-        var curCanvas;
-        if (args.pyramid) {
-            curCanvas = args.pyramid[i];
-            if (
-                Math.abs(curCanvas.width - width) > 1e-3 ||
-                Math.abs(curCanvas.height - height) > 1e-3
-            )
-                throw new Error("Adding SSV: Canvas sizes do not match.");
-        } else {
-            curCanvas = new Canvas(
-                "ssv" + (this.ssvs.length - 1) + "_" + "level" + i,
-                width,
-                height
-            );
-            this.addCanvas(curCanvas);
-        }
-        curPyramid.push(curCanvas);
-
-        // create one layer
-        var curLayer = new Layer(transform, false);
-        curCanvas.addLayer(curLayer);
-
-        // set fetching scheme
-        if (ssv.clusterMode == "contour" || ssv.clusterMode == "heatmap")
-            curLayer.setFetchingScheme("dbox", false);
-        //curLayer.setFetchingScheme("tiling");
-
-        // set ssv ID
-        curLayer.setIndexerType("SSVInMemoryIndexer");
-        //curLayer.setIndexerType("SSVCitusIndexer");
-        curLayer.setSSVId(this.ssvs.length - 1 + "_" + i);
-
-        // dummy placement
-        curLayer.addPlacement({
-            centroid_x: "con:0",
-            centroid_y: "con:0",
-            width: "con:0",
-            height: "con:0"
-        });
-
-        // construct rendering function
-        curLayer.addRenderingFunc(ssv.getLayerRenderer());
-
-        // tooltips
-        curLayer.addTooltip(ssv.tooltipColumns, ssv.tooltipAliases);
-
-        // map layer
-        if (ssv.mapBackground) {
-            var mapLayer = new Layer(
-                require("./Transform").defaultEmptyTransform,
-                false
-            );
-            curCanvas.addLayer(mapLayer);
-            mapLayer.addRenderingFunc(ssv.getMapRenderer());
-            mapLayer.addPlacement({
-                centroid_x: "con:0",
-                centroid_y: "con:0",
-                width: "con:0",
-                height: "con:0"
-            });
-            mapLayer.setFetchingScheme("dbox", false);
-            mapLayer.setSSVId(this.ssvs.length - 1 + "_" + i);
-        }
-
-        // add static legend layer
-        var staticLayer = new Layer(null, true);
-        curCanvas.addLayer(staticLayer);
-        staticLayer.addRenderingFunc(ssv.getLegendRenderer());
-        staticLayer.setSSVId(this.ssvs.length - 1 + "_" + i);
-
-        // axes
-        if (ssv.axis) {
-            curCanvas.addAxes(
-                ssv.getAxesRenderer(i),
-                "ssv_" + (this.ssvs.length - 1)
-            );
-        }
-    }
-
-    // literal zooms
-    for (var i = 0; i + 1 < ssv.numLevels; i++) {
-        var hasLiteralZoomIn = false;
-        var hasLiteralZoomOut = false;
-        for (var j = 0; j < this.jumps.length; j++) {
-            if (
-                this.jumps[j].sourceId == curPyramid[i].id &&
-                this.jumps[j].type == "literal_zoom_in"
-            ) {
-                if (this.jumps[j].destId != curPyramid[i + 1].id)
-                    throw new Error(
-                        "Adding SSV: malformed literal zoom pyramid."
-                    );
-                hasLiteralZoomIn = true;
-            }
-            if (
-                this.jumps[j].sourceId == curPyramid[i + 1].id &&
-                this.jumps[j].type == "literal_zoom_out"
-            ) {
-                if (this.jumps[j].destId != curPyramid[i].id)
-                    throw new Error(
-                        "Adding SSV: malformed literal zoom pyramid."
-                    );
-                hasLiteralZoomOut = true;
-            }
-        }
-        if (!hasLiteralZoomIn)
-            this.addJump(
-                new Jump(curPyramid[i], curPyramid[i + 1], "literal_zoom_in")
-            );
-        if (!hasLiteralZoomOut)
-            this.addJump(
-                new Jump(curPyramid[i + 1], curPyramid[i], "literal_zoom_out")
-            );
-    }
-
-    // create a new view if not specified
-    if (!args.view) {
-        var viewId = "ssv" + (this.ssvs.length - 1);
-        var view = new View(viewId, ssv.topLevelWidth, ssv.topLevelHeight);
-        this.addView(view);
-        // initialize view
-        this.setInitialStates(view, curPyramid[0], 0, 0);
-    } else if (!(args.view instanceof View))
-        throw new Error("Adding SSV: view must be a View object");
-
-    return {pyramid: curPyramid, view: args.view ? args.view : view};
-}
-
-/**
- * Add a USMap template object to a project
- * @param usmap a USMap object
- * @param args an dictionary that contains customization parameters, see doc
- */
-function addUSMap(usmap, args) {
-    if (args == null) args = {};
-    var numCanvas = "county" in usmap ? 2 : 1;
-    if ("pyramid" in args && args.pyramid.length != numCanvas)
-        throw new Error(
-            "Adding USMap: args.pyramid does not have matching number of canvases"
-        );
-
-    // add to project
-    this.usmaps.push(usmap);
-
-    // rendering params
-    var rpKey = "usmap_" + (this.usmaps.length - 1);
-    var rpDict = {};
-    rpDict[rpKey] = usmap.params;
-    this.addRenderingParams(rpDict);
-
-    // ================== state map canvas ===================
-    var canvases = [];
-    var stateMapCanvas;
-    if ("pyramid" in args) stateMapCanvas = args.pyramid[0];
-    else {
-        stateMapCanvas = new Canvas(
-            "usmap" + (this.usmaps.length - 1) + "_" + "state",
-            usmap.stateMapWidth,
-            usmap.stateMapHeight
-        );
-        this.addCanvas(stateMapCanvas);
-    }
-    if (
-        stateMapCanvas.w != usmap.stateMapWidth ||
-        stateMapCanvas.h != usmap.stateMapHeight
-    )
-        throw new Error("Adding USMap: state canvas sizes do not match");
-
-    // static legends layer
-    var stateMapLegendLayer = new Layer(null, true);
-    stateMapCanvas.addLayer(stateMapLegendLayer);
-    stateMapLegendLayer.addRenderingFunc(
-        usmap.getUSMapRenderer("stateMapLegendRendering")
-    );
-    stateMapLegendLayer.setUSMapId(this.usmaps.length - 1 + "_" + 0);
-
-    // state boundary layer
-    var stateMapTransform = new Transform(
-        `SELECT name, ${usmap.stateRateCol}, geomstr 
-         FROM ${usmap.stateTable}`,
-        usmap.db,
-        usmap.getUSMapTransformFunc("stateMapTransform"),
-        ["bbox_x", "bbox_y", "name", "rate", "geomstr"],
-        true
-    );
-    var stateBoundaryLayer = new Layer(stateMapTransform, false);
-    stateMapCanvas.addLayer(stateBoundaryLayer);
-    stateBoundaryLayer.addPlacement({
-        centroid_x: "col:bbox_x",
-        centroid_y: "col:bbox_y",
-        width: `con:${usmap.stateMapWidth / usmap.zoomFactor}`,
-        height: `con:${usmap.stateMapWidth / usmap.zoomFactor}`
-    });
-    stateBoundaryLayer.addRenderingFunc(
-        usmap.getUSMapRenderer("stateMapRendering")
-    );
-    stateBoundaryLayer.addTooltip(
-        ["name", "rate"],
-        ["State", usmap.tooltipAlias]
-    );
-    stateBoundaryLayer.setUSMapId(this.usmaps.length - 1 + "_" + 0);
-
-    // add to canvases (return)
-    canvases.push(stateMapCanvas);
-
-    // ==========  Views ===============
-    if (!("view" in args)) {
-        var view = new View(
-            "usmap" + (this.usmaps.length - 1),
-            usmap.stateMapWidth,
-            usmap.stateMapHeight
-        );
-        this.addView(view);
-        this.setInitialStates(view, stateMapCanvas, 0, 0);
-    } else if (!(args.view instanceof View)) {
-        throw new Error("Adding USMap: view must be a View object");
-    }
-
-    // ================== county map canvas ===================
-    if ("countyTable" in usmap) {
-        var countyMapCanvas;
-        if ("pyramid" in args) countyMapCanvas = args.pyramid[1];
-        else {
-            countyMapCanvas = new Canvas(
-                "usmap" + (this.usmaps.length - 1) + "_" + "county",
-                usmap.stateMapWidth * usmap.zoomFactor,
-                usmap.stateMapHeight * usmap.zoomFactor
-            );
-            this.addCanvas(countyMapCanvas);
-        }
-        if (
-            countyMapCanvas.w != usmap.stateMapWidth * usmap.zoomFactor ||
-            countyMapCanvas.h != usmap.stateMapHeight * usmap.zoomFactor
-        )
-            throw new Error("Adding USMap: county canvas sizes do not match");
-
-        // static legends layer
-        var countyMapLegendLayer = new Layer(null, true);
-        countyMapCanvas.addLayer(countyMapLegendLayer);
-        countyMapLegendLayer.addRenderingFunc(
-            usmap.getUSMapRenderer("countyMapLegendRendering")
-        );
-        countyMapLegendLayer.setUSMapId(this.usmaps.length - 1 + "_" + 1);
-
-        // thick state boundary layer
-        var countyMapStateBoundaryTransform = new Transform(
-            `SELECT geomstr FROM ${usmap.stateTable}`,
-            usmap.db,
-            usmap.getUSMapTransformFunc("countyMapStateBoundaryTransform"),
-            ["bbox_x", "bbox_y", "bbox_w", "bbox_h", "geomstr"],
-            true
-        );
-        var countyMapStateBoundaryLayer = new Layer(
-            countyMapStateBoundaryTransform,
-            false
-        );
-        countyMapCanvas.addLayer(countyMapStateBoundaryLayer);
-        countyMapStateBoundaryLayer.addPlacement({
-            centroid_x: "col:bbox_x",
-            centroid_y: "col:bbox_y",
-            width: "col:bbox_w",
-            height: "col:bbox_h"
-        });
-        countyMapStateBoundaryLayer.addRenderingFunc(
-            usmap.getUSMapRenderer("countyMapStateBoundaryRendering")
-        );
-        countyMapStateBoundaryLayer.setUSMapId(
-            this.usmaps.length - 1 + "_" + 1
-        );
-
-        // county boundary layer
-        var countyMapTransform = new Transform(
-            `SELECT name, ${usmap.countyRateCol}, geomstr
-        FROM ${usmap.countyTable};`,
-            usmap.db,
-            usmap.getUSMapTransformFunc("countyMapTransform"),
-            ["bbox_x", "bbox_y", "bbox_w", "bbox_h", "name", "rate", "geomstr"],
-            true
-        );
-        var countyBoundaryLayer = new Layer(countyMapTransform, false);
-        countyMapCanvas.addLayer(countyBoundaryLayer);
-        countyBoundaryLayer.addPlacement({
-            centroid_x: "col:bbox_x",
-            centroid_y: "col:bbox_y",
-            width: "col:bbox_w",
-            height: "col:bbox_h"
-        });
-        countyBoundaryLayer.addRenderingFunc(
-            usmap.getUSMapRenderer("countyMapRendering")
-        );
-        countyBoundaryLayer.addTooltip(
-            ["name", "rate"],
-            ["County", usmap.tooltipAlias]
-        );
-        countyBoundaryLayer.setUSMapId(this.usmaps.length - 1 + "_" + 1);
-
-        // add to canvases (return)
-        canvases.push(countyMapCanvas);
-
-        // =============== jump ===============
-        if (usmap.zoomType == "literal") {
-            this.addJump(
-                new Jump(stateMapCanvas, countyMapCanvas, "literal_zoom_in")
-            );
-            this.addJump(
-                new Jump(countyMapCanvas, stateMapCanvas, "literal_zoom_out")
-            );
-        } else if (usmap.zoomType == "jump") {
-            var selector = new Function(
-                "row",
-                "args",
-                `return args.layerId = ${stateMapCanvas.layers.length - 1}`
-            );
-            var newPredicates = function() {
-                return {};
-            };
-            var newViewportBody = function(row, args) {
-                var zoomFactor = REPLACE_ME_zoomfactor;
-                var vpW = args.viewportW;
-                var vpH = args.viewportH;
-                return {
-                    constant: [
-                        row.bbox_x * zoomFactor - vpW / 2,
-                        row.bbox_y * zoomFactor - vpH / 2
-                    ]
-                };
-            };
-            var newViewport = new Function(
-                "row",
-                "args",
-                getBodyStringOfFunction(newViewportBody).replace(
-                    /REPLACE_ME_zoomfactor/g,
-                    usmap.zoomFactor
-                )
-            );
-            var jumpName = function(row) {
-                return "County map of " + row.name;
-            };
-            this.addJump(
-                new Jump(
-                    stateMapCanvas,
-                    countyMapCanvas,
-                    "geometric_semantic_zoom",
-                    {
-                        selector: selector,
-                        viewport: newViewport,
-                        predicates: newPredicates,
-                        name: jumpName
-                    }
-                )
-            );
-        }
-    }
-
-    return {pyramid: canvases, view: args.view ? args.view : view};
-}
-
-function addPie(pie, args) {
-    if (args == null) args = {};
-
-    // add to project
-    this.pies.push(pie);
-
-    // construct canvas
-    var pieCanvas;
-    if ("canvas" in args) pieCanvas = args.canvas;
-    else {
-        pieCanvas = new Canvas(
-            "pie" + (this.pies.length - 1),
-            pie.width,
-            pie.height
-        );
-        this.addCanvas(pieCanvas);
-    }
-    if (pieCanvas.w != pie.width || pieCanvas.h != pie.height)
-        throw new Error("Adding Pie: canvas sizes do not match.");
-
-    // add rendering params
-    var rpKey = "pie_" + (this.pies.length - 1);
-    var rpDict = {};
-    rpDict[rpKey] = {
-        dimensions: pie.query.dimensions,
-        innerRadius: 70,
-        outerRadius: pie.radius,
-        cornerRadius: 5,
-        padAngle: 0.01,
-        colorScheme: pie.colorScheme,
-        transition: pie.transition,
-        legendTitle: pie.legend.title,
-        legendDomain: pie.legend.domain
-    };
-    this.addRenderingParams(rpDict);
-
-    // construct query
-    // SELECT columns are from measureCol & pie.query.dimensions
-    // merge them and then dedup
-    var query = "SELECT " + pie.query.dimensions.join(", ");
-    query += (pie.query.dimensions.length ? ", " : "") + pie.query.measure;
-    query += " FROM " + pie.query.table + " GROUP BY ";
-    query += pie.query.dimensions.join(", ");
-    var pieTransform = new Transform(
-        query,
-        pie.db,
-        "",
-        pie.query.dimensions.concat(["kyrixAggValue"]),
-        true,
-        pie.query.dimensions.concat(pie.query.sampleFields)
-    );
-
-    // construct pie layer
-    var pieLayer = new Layer(pieTransform, true);
-    pieCanvas.addLayer(pieLayer);
-    pieLayer.addRenderingFunc(pie.getPieRenderer());
-    pieLayer.addTooltip(pie.tooltip.columns, pie.tooltip.aliases);
-    pieLayer.setIndexerType("StaticAggregationIndexer");
-    pieLayer.setPieId(this.pies.length - 1 + "_" + 0);
-
-    // construct queries for the dummy sample layer
-    // construct the dummy sample layer
-    query =
-        "SELECT " +
-        pie.query.sampleFields.join(", ") +
-        (pie.query.sampleFields.length ? ", " : "") +
-        pie.query.dimensions.join(", ") +
-        (pie.query.measureCol === "*"
-            ? ""
-            : (pie.query.sampleFields.length || pie.query.dimensions.length
-                  ? ", "
-                  : "") + pie.query.measureCol) +
-        " FROM " +
-        pie.query.table;
-
-    // construct sample layer
-    var sampleTransform = new Transform(query, pie.db, "", [], true);
-    var sampleLayer = new Layer(sampleTransform, true);
-    pieCanvas.addLayer(sampleLayer);
-    sampleLayer.addRenderingFunc(function() {});
-    sampleLayer.setIndexerType("StaticAggregationIndexer");
-
-    // add stylesheet
-    this.addStyles(__dirname + "/template-api/css/pie.css");
-
-    // view
-    if (!("view" in args)) {
-        var view = new View(
-            "pie" + (this.pies.length - 1),
-            pie.width,
-            pie.height
-        );
-        this.addView(view);
-        this.setInitialStates(view, pieCanvas, 0, 0);
-    } else if (!(args.view instanceof View)) {
-        throw new Error("Adding Pie: view must be a View object");
-    }
-
-    return {canvas: pieCanvas, view: args.view ? args.view : view};
 }
 
 // Add a rendering parameter object
